requirements.txt
asgiref==3.8.1
babel==2.17.0
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
Django==5.2.1
django-jazzmin==3.0.1
django-money==3.5.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
py-moneyed==3.0
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.0
requests==2.32.3
setuptools==80.3.1
sqlparse==0.5.3
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
